refactor: migrate PlayerContext imports
